Work in progress: Tried using TPT hierarchy for doctor and pacient

+++ ReservationsAPI/ReservationsAPI/DAL/Entities/Appointment.cs
@@ -3,8 +3,8 @@ namespace ReservationsAPI.DAL.Entities
 {
     public class Appointment
     {
-        public long PacientId { get; set; }
-        public long DoctorId { get; set; }
+        public int PacientId { get; set; }
+        public int DoctorId { get; set; }
         public long ProcedureId { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
