+++ ReservationsAPI/ReservationsAPI/DAL/Entities/VacationDay.cs
@@ -4,7 +4,7 @@ namespace ReservationsAPI.DAL.Entities
     public class VacationDay
     {
         public long Id { get; set; }
-        public long DoctorId { get; set; }
+        public int DoctorId { get; set; }
         public DateTime Date { get; set; }
 
         public virtual Doctor Doctor { get; set; }
